+++ packages/cli-plugin-fc/src/index.ts
@@ -33,10 +33,13 @@ export class AliyunFCPlugin extends BasePlugin {
     'package:generateEntry': async () => {
       this.core.cli.log('Generate entry file...');
       const { version } = findMidwayVersion(this.servicePath);
-      this.setGlobalDependencies(
-        '@midwayjs/serverless-fc-starter',
-        version.major || '*'
-      );
+      const specificStarterName = this.core.service?.provider?.starter;
+      if (!specificStarterName) {
+        this.setGlobalDependencies(
+          '@midwayjs/serverless-fc-starter',
+          version.major || '*'
+        );
+      }
       const preloadFile = this.getStore('preloadFile', 'global');
 
       writeWrapper({
@@ -44,6 +47,7 @@ export class AliyunFCPlugin extends BasePlugin {
         service: this.core.service,
         distDir: this.midwayBuildPath,
         starter: '@midwayjs/serverless-fc-starter',
+        specificStarterName,
         preloadFile,
       });
     },