+++ packages/cli-plugin-scf/src/index.ts
@@ -34,14 +34,19 @@ export class TencentSCFPlugin extends BasePlugin {
     'package:generateEntry': async () => {
       this.core.cli.log('Generate entry file...');
       const { version } = findMidwayVersion(this.servicePath);
-      this.setGlobalDependencies(
-        '@midwayjs/serverless-scf-starter',
-        version.major || '*'
-      );
+      const specificStarterName = this.core.service?.provider?.starter;
+      if (!specificStarterName) {
+        this.setGlobalDependencies(
+          '@midwayjs/serverless-scf-starter',
+          version.major || '*'
+        );
+      }
+
       writeWrapper({
         baseDir: this.servicePath,
         service: this.core.service,
         distDir: this.midwayBuildPath,
+        specificStarterName: this.core.service?.provider?.specificStarterName,
         starter: '@midwayjs/serverless-scf-starter',
       });
     },