+++ packages/cli-plugin-fc/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.8-beta.4](https://github.com/midwayjs/cli/compare/v1.3.8-beta.3...v1.3.8-beta.4) (2022-07-16)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-fc
+
+
+
+
+
+## [1.3.8-beta.3](https://github.com/midwayjs/cli/compare/v1.3.8-beta.2...v1.3.8-beta.3) (2022-07-16)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-fc
+
+
+
+
+
+## [1.3.8-beta.2](https://github.com/midwayjs/cli/compare/v1.3.8-beta.1...v1.3.8-beta.2) (2022-07-12)
+
+
+### Bug Fixes
+
+* v3 starter ([1af272c](https://github.com/midwayjs/cli/commit/1af272cb9efaf32681604ceaa112ce746c8ca0cb))
+
+
+
+
+
+## [1.3.8-beta.1](https://github.com/midwayjs/cli/compare/v1.3.7...v1.3.8-beta.1) (2022-07-12)
+
+
+### Bug Fixes
+
+* new faas starter ([55d40a4](https://github.com/midwayjs/cli/commit/55d40a4d200be4e3761e294452c7cd77778ebbb2))
+
+
+### Features
+
+* new faas & check ([9eb6133](https://github.com/midwayjs/cli/commit/9eb61337ca000b201b70c16f85ab6e9f1f74db4a))
+
+
+
+
+
 ## [1.3.7](https://github.com/midwayjs/cli/compare/v1.3.6...v1.3.7) (2022-07-06)