+++ packages/serverless-spec-builder/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.8-beta.4](https://github.com/midwayjs/cli/compare/v1.3.8-beta.3...v1.3.8-beta.4) (2022-07-16)
+
+
+### Bug Fixes
+
+* aggregation handler name ([7ed1d48](https://github.com/midwayjs/cli/commit/7ed1d483b86cf24c619026ffa9d1e02f861cea39))
+
+
+
+
+
+## [1.3.8-beta.3](https://github.com/midwayjs/cli/compare/v1.3.8-beta.2...v1.3.8-beta.3) (2022-07-16)
+
+
+### Bug Fixes
+
+* aggregation wrapper ([74b4560](https://github.com/midwayjs/cli/commit/74b4560f8d1903d2c33e50e28dd16e45219c4617))
+
+
+
+
+
+## [1.3.8-beta.2](https://github.com/midwayjs/cli/compare/v1.3.8-beta.1...v1.3.8-beta.2) (2022-07-12)
+
+
+### Bug Fixes
+
+* v3 starter ([1af272c](https://github.com/midwayjs/cli/commit/1af272cb9efaf32681604ceaa112ce746c8ca0cb))
+
+
+
+
+
+## [1.3.8-beta.1](https://github.com/midwayjs/cli/compare/v1.3.7...v1.3.8-beta.1) (2022-07-12)
+
+
+### Features
+
+* new faas & check ([9eb6133](https://github.com/midwayjs/cli/commit/9eb61337ca000b201b70c16f85ab6e9f1f74db4a))
+
+
+
+
+
 ## [1.3.6](https://github.com/midwayjs/cli/compare/v1.3.5...v1.3.6) (2022-06-02)